Behaviour tests and State tests implemented in one file just separately
